pyproject.toml
[project]
name = "brain-tumor-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyter>=1.1.1",
    "mlflow>=2.20.3",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.0.1",
    "roboflow",
    "seaborn>=0.13.2",
    "tensorflow[and-cuda]>=2.18.0",
]

[tool.uv.sources]
imgaug = { git = "https://github.com/aleju/imgaug.git" }
